algorithm/tracking/requirements.txt
munkres
pyproj
filterpy
ortools
torch
numpy
numba
scipy
opencv-python
torchvision
pillow
motmetrics
open3d